+++ extensions/ql-vscode/src/query-server/index.ts
@@ -1,4 +1,3 @@
-export * from "./new-query-runner";
 export * from "./query-runner";
 export * from "./query-server-client";
 export * from "./run-queries";
